requirements.txt

aioredis==2.0.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
databases==0.8.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.104.1
fastapi-limiter==0.1.5
h11==0.14.0
idna==3.6
jose==1.0.0
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0